+++ src/generator/annotations.ts
@@ -1,4 +1,5 @@
 export const DTO_IGNORE_MODEL = /@DtoIgnoreModel/;
+export const DTO_SINGULAR_MODEL = /@DtoSingularModel/;
 export const DTO_READ_ONLY = /@DtoReadOnly/;
 export const DTO_ENTITY_HIDDEN = /@DtoEntityHidden/;
 export const DTO_CREATE_OPTIONAL = /@DtoCreateOptional/;